requirements.txt
absl-py==1.0.0
asgiref==3.5.0
astunparse==1.6.3
backports.zoneinfo==0.2.1
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.4
django-cors-headers==3.11.0
django-import-export==2.8.0
djangorestframework==3.13.1
drf-yasg==1.20.0
et-xmlfile==1.1.0
flatbuffers==2.0
fonttools==4.33.3
gast==0.5.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.1
Markdown==3.3.6
MarkupPy==1.14
MarkupSafe==2.1.1
matplotlib==3.5.1
numpy==1.22.3
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
tablib==3.2.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
typing-extensions==4.2.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.8.0
